refactor(proposals): extract activity, prompt and json tabs from edit page
The proposal edit page drops from 3,196 to 2,358 lines: the three
remaining fully-inline tabs move to self-contained components under
components/panel/proposal/. ProposalJsonTab owns export/import (with its
own ConfirmModal instance) and emits 'applied' so the page rehydrates the
General form; ProposalPromptTab owns both seller/technical prompt
composables and loads them on mount; ProposalActivityTab owns the log
form and timeline formatting. JSON stats math moves to
utils/proposalJsonStats.js, shared by the page's technical sub-tab and
the new tab. All data-testids preserved.
